requirements.txt
sentence-transformers>=2.2.0
numpy>=1.21.0
rich>=10.0.0
anthropic>=0.5.0
python-dotenv>=0.19.0